Fixing bug 568284 (nullref on Clean command). (#14)
* Fixing bug 568284 (nullref on Clean command).

It's a bit ugly but I think it's OK for preview 3. Not sure if we should propagate the update failure out and fail the entire Clean operation (may have cleaned some but not all libraries - success or failure?).

We should test the FileSystem provider though as well. The code says Files should be required, not sure if it's enforced.

* Addressing code review feedback.

* Addressing code review feedback.