Use a fake WebRequestHandler for calls that JsDelivrCatalog makes directly
This change is almost entirely to the tests, trying to make them more accurate and more reliable.

JsDelivrCatalog also relies on NpmPackageSearch, which is still hardcoded to a live WebRequestHandler.  That'll be yet another refactoring pass to be able to fake it out.

Added a couple TODOs for missing scenarios that will require product code changes.