Ensure completions in Add wizard are sorted (#647)
There isn't a guarantee that the data from providers is sorted in the
desired order (e.g. for versions, sorted descending).  This change
ensures we sort the completion items before showing them in the search
box flyout.

In the JSON completion scenario (which is similar), each completion item
is instead represented by a CompletionEntry implementation which
implements IComparable<T> for that completion type.  The editor
completion stack handles the sorting for us already.

Addresses #531.